requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
appdirs==1.4.4
atomicwrites==1.4.1
attrs==25.3.0
black==22.12.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
filetype==1.2.0
h11==0.14.0
har2case==0.3.1
httprunner==3.1.11
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
jmespath==0.9.5
jsonpath==0.82.2
jsonschema==3.2.0
loguru==0.4.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==0.13.1
py==1.11.0
pycparser==2.22
pydantic==1.8.2
pyrsistent==0.20.0
PySocks==1.7.1
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==3.1.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==0.9.1
selenium==4.27.1
sentry-sdk==0.14.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.2.1
trio==0.27.0
trio-websocket==0.12.2
typed-ast==1.5.5
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
win32_setctime==1.2.0
wsproto==1.2.0
